Ensure outbound links open in new tab
Otherwise visiting the schedule and clicking back will break Element from loading *if* the user has clicked one of the hash urls to get to a section on the page. Also loading Element takes time, so a new tab is friendlier as well.

Additionally add `rel="noopener noreferrer"` to the new links as good practice.